elvis_bank

+++ src/Exception_student_management/Mark.java
@@ -3,16 +3,16 @@ package Exception_student_management;
 public class Mark {
     Students s;
     Subject sj;
-    float m;
+    Double m;
     int bonus;
-    Mark(Students _s, Subject _sj, float _m, int _bonus){
+    Mark(Students _s, Subject _sj, Double _m, int _bonus){
         s = _s;
         sj = _sj;
         m = _m;
         bonus = _bonus;
     }
-    public float score() {
-        float score;
+    public Double score() {
+        Double score;
         score = m * sj.w;
         return score;
     }
